Update 0.2
Добавлены плейлисты.
В связи с этим добавлены:
- окно добавления плейлиста
- база данных с плейлистами
Переработаны:
- окна добавления и изменения шаблона
- база данных шаблонов

+++ Six Screens Controller/App.xaml.cs
@@ -103,7 +103,8 @@ namespace Six_Screens_Controller
         {
             try
             {
-                process.Kill();
+                if (config.Python != "" && config.Server != "")
+                    process.Kill();
             }
             catch (Exception ex)
             {
